Add config for overriding username
This is for folks who have quay.io usernames different from their local OS username. By default Quarkus uses the value of Java's `user.name` property which is by default the user's local hostname (e.g. the value of `whoami` on Linux). If the two aren't the same, Quarkus will fail when it tries to push the container image to the wrong Quay repository spec.